main: handle ExtensionCommandError better
Extension commands can fail to load early in the process, when we load
the extension command information from the file system. We are not
currently handling this in a nice way, resulting in tracebacks instead
of sensible error messages.

Fix this by catching ExtensionCommandError and using similar error
handling as we are already doing from within run_command().

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>