main.py: rely on WestCommand to do output
Remove most west.log API calls by delegating the output to the
WestCommand instance we are using.

Most cases are handled by saving the output we want to print in an
instance attribute 'queued_io', which gets flushed by the command
itself as a pre-run hook.

Any I/O we need to do in exception handling blocks following the
execution of the command can just use the command directly.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>