treewide: don't call Project.git() with strings
Hypothesis: special characters when doing things like commit 'peeling'
in refs like "v3.0.0^{commit}" are being treated as metacharacters by
the windows cmd.exe on some people's machines:

   You must use quotation marks around the following special
   characters: & < > [ ] | { } ^ = ; ! ' + , ` ~ [white space].

   https://docs.microsoft.com/en-us/windows-server/administration/windows-commands/cmd#syntax

If so, this should be a fix, since we'll no longer be using
shlex.split() in Project.git() for git() calls within west itself.
We definitely do include ^, {, and } in our command strings, and
shlex.split() is a POSIX-specific split, which may not be doing the
right thing on Windows, despite it passing CI.

Some Zephyr users are reporting problems that appear suspiciously like
this being the case:

https://github.com/zephyrproject-rtos/zephyr/discussions/48837

Unclear why this doesn't reproduce in our own testing.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>