tests: work around new git submodule behavior
This is a work-around for:

https://github.blog/2022-10-18-git-security-vulnerabilities-announced/#cve-2022-39253

Our test cases are failing since they are using 'git submodule add' on
a URL which is a file. Recent versions of git are now rejecting this
by default due to the above CVE. We can override this default behavior
on a per-git-command basis by overriding the configuration option for
the duration of that process. There is no risk here since the
associated repositories are not maliciously crafted.

This work-around follows a suggestion here:

https://bugs.launchpad.net/ubuntu/+source/git/+bug/1993586/comments/3

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>