version: v1.0.0a1
West has settled down to the point that it is time to declare its
internal APIs stable. We've been promising since the beginning that
API instability was present until we cut v1.0, so moving to this
version is a signal to users that the program will bump its major
version when any big changes happen from now on.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>