compare: add command
This command prints information about manifest-rev and "git status"
output for a project if (and only if) at least one of the following is
true:

1. its checked out commit is different than the commit checked
   out by the most recent successful "west update"
2. its working tree is not clean (i.e. there are local uncommitted
   changes)
3. it has a local checked out branch (unless overridden via
   the compare.ignore-branches config option or the --ignore-branches
   command line option)

It does something similar for the manifest repository.

The purpose of this is to allow people who are developing on one or
more projects to quickly see if they have anything interesting going
on, or if their projects match the manifest-rev versions.

Fixes: #642
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>