manifest: initial import filtering implementation
This feature provides a way to prevent west from performing an import
depending on a condition.

There are a variety of use cases for this:

- CI system maintainers want to be able to define an 'import' that
  contains repositories that contain test cases and which should be
  easy for developers to fetch with 'west update', but which are not
  distributed to users by default (by 'turning off' the import
  somehow)

- Similarly, but for projects which provide optional features that can
  be safely ignored unless the user takes an explicit action to enable
  the import. The main use case here isn't limiting visibility of
  the feature, but rather limiting the storage size and network
  bandwidth required to initialize a basic workspace

Provide this feature by extending the 'import: <map>' syntax with a
new key, 'if':

  projects:
    - name: foo
      import:
        if:
          <condition>

Where the value of <condition> is itself a map, so that we can add
features to it over time. In this commit, we add the infrastructure
for the 'if' key and one simple way to write <condition>:

  import:
    if:
      config-enabled: foo.bar

Here, the import will be resolved if and only if the west
configuration option 'foo.bar' is defined and set to a true boolean
value (see the python configparser docs for all the ways to write a
boolean option; "west config foo.bar true" is one way to set this up).

The configuration option can be set in any configuration file: local,
global, or system. This makes it easier for users who will want the
import performed but regularly set up workspaces in various places to
set it up permanently, without having to enable foo.bar every time
they set up a workspace, before the first "west update".

Design notes for the record:

- I was originally trying to do something like "import: if:
  group-enabled: bsim", but that runs into possibilities for
  self-contradiction, because a project that is imported later than
  you can disable a group, so you don't really know if a group is
  enabled or not until you are done parsing the manifest.

- It thus made sense to me to define a syntax for conditional
  imports that doesn't rely on groups at all. Configuration options
  are independent of groups and are set once and for all by the time
  we are loading the manifest, so it seemed like a good place to
  start.

- People are still rapidly ideating new ways to do import
  filtering for different use cases, so a one-size-fits-all solution
  seems unlikely. Making a flexible syntax that we can extend later
  therefore seemed more future-proof. In the future, we can do things
  like:

    import:
      if:
        env: SOME_ENV_VAR=1

  or:

    import:
      if:
        anyOf:
          config-enabled: foo.bar
          env: SOME_ENV_VAR=1

  etc. There seems to be plenty of room to grow.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>