manifest: initial manifest.project-filter project name validation
In order to support the manifest.project-filter configuration option,
we need to forbid certain characters from occurring in project names.

Making that happen is a backwards incompatible change, however, so
phase it in by emitting a warning.

We will make it a hard error if manifest.project-filter is set later
on. That way, users will only see an error if they are opting into the
latest west for now, and other users will get a warning and have time
to migrate their projects.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>