app: re-work logging yet again
Commit 92c18ac55acd631c11f9d4392556d3b7075a0bec
("main: move verbose manifest logging to project.py")
did not achieve its purpose.

We are loading the manifest long before we call the setup_logging()
methods in each of the project.py classes, so any messages from the
manifest class have long been discarded by the time that we enable
them.

Now that we have EarlyArgs in main.py, we can use that to centralize
logging and enforce the following:

- warnings and above for west APIs are emitted by default
- info and above emitted with west -v
- debug and above with west -vv or higher

Fixes: #665
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>