Revert "tests: fix test_group_filter_imports()"
This reverts commit d0e6b9a54b4d176cdf474b7ae7a05e8d547cf17c.

West was right, the docs are wrong/inconsistent.

The commit I'm reverting was following the west documentation it
referenced correctly, but the documentation itself was wrong (and in
fact was inconsistent with earlier documentation on the same page).

The basis for making commit d0e6b9a54 was that the docs currently say:

    In other words, let:

    - the submanifest resolved from self-import have group filter self-filter
    - the top-level manifest file have group filter top-filter
    - the submanifests resolved from import-1 through import-N have
      group filters filter-1 through filter-N respectively

    The final resolved group-filter value is then filter1 + filter-2 +
    ... + filter-N + top-filter + self-filter, where + here refers to
    list concatenation.

But earlier in the same section, they also say:

    The resolved manifest has a group-filter value which is the result
    of concatenating the group-filter values in the top-level manifest
    and any imported manifests.

    Manifest files which appear earlier in the import order have
    higher precedence and are therefore concatenated later into the
    final group-filter.

Where the "import order" is defined higher up as:

    Importing is done in this order:

    1. Manifests from self-import are imported first.
    2. The top-level manifest file’s definitions are handled next.
    3. Manifests from import-1, …, import-N, are imported in that order.

This means that import-1 happens **earlier** than import-N, and should
therefore have **higher** precedence when it comes to computing the
final group filter. But putting import-N **later** in the concanated
group filter actually means import-1 has **lower** precedence. So the
docs are inconsistent.

To fix the docs so they are consistent and match west's current
behavior, we should replace this:

    The final resolved group-filter value is then filter1 + filter-2 + ...
    + filter-N + top-filter + self-filter, where + here refers to list
    concatenation.

with this:

    The final resolved group-filter value is then filter-N + ... +
    filter-1 + top-filter + self-filter, where + here refers to list
    concatenation.

That is:

- docs currently go from: filter-1 to filter-N
- they should go from:    filter-N to filter-1.

Fixes: #663
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>