commands: Show less generic error when no manifest is loaded
If loading the manifest fails, running a command which needs to read the
manifest could result in this being shown:

   FATAL ERROR: can't run west manifest; it requires the manifest, which
   was not available. Try "west manifest --validate" to debug.

In this case, trying "west manifest --validate" would also display the
same error, which is not very helpful.

Instead of that, west can print the deferred error from load_manifest(),
which contains a more specific message. Rework the WestCommand.manifest
property handling by adding a dummy _ManifestRequired exception. It is
caught by run_builtin() in main.py, which can access the stored error.

Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>