commands: fix some docstrings
When interpreted as restructured text, these are causing errors
because the '**' is interpreted as the start of a bold section, which
never ends. This causes build failures in the zephyr docs, which do
treat these as RST.

Wrap them in `` to make the whole thing a literal block to fix this
issue.

Signed-off-by: Martí Bolívar <mbolivar@amperecomputing.com>