west: Add untracked files and folders to west status output
Add untracked files and folder to west status output.
This helps to keep the west checkout clean by alerting the user
of folders that are no longer tracked by west.
This is useful for detecting when a module has been removed, or moved.
and is no longer updated by west update.

Add the ability to ignore folders and files through a comma separated
list in the west config file.
Example:
west config status.ignore .vscode/,optional/

Fixes: #622

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>