main: Fix unbound 'spec' variable when loading extension commands
The 'spec' loop variable was not captured, and this would result in
the lambda function only evaluating to the last value in the loop.

The issue could occur when loading extension commands from projects,
where the name conflicts with a builtin command or an already loaded
extension command.

Discovered with flake8-bugbear, see
https://docs.astral.sh/ruff/rules/function-uses-loop-variable/

Fix function-uses-loop-variable (B023)

Signed-off-by: Pieter De Gendt <pieter.degendt@basalte.be>