tests: manifest: empty name-allowlist not the same as None
When using the west manifest import feature, name-allowlist
should have a slightly different interpretation when it is

a) unspecified (all imports are allowed)
b) specified as empty (no imports of child project are allowed)
c) specified as non-empty (only specified imports allowed)

The a) and c) cases already have tests.

This change adds a test to cover the b) case.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>